requirements.txt
Django==5.2.6
asgiref~=3.8
sqlparse~=0.5.1
tzdata>=2024.1 ; sys_platform == "win32"
django-crispy-forms==2.3
crispy-bootstrap5==2024.2
python-dotenv>=1.0.1    
Pillow>=10.4          
# beautifulsoup4>=4.12.3
# whitenoise>=6.7
# gunicorn>=23.0
# dj-database-url>=2.2
# psycopg[binary]>=3.2

